Update nixpkgs
Singular up to 4.4.0 would unconditionally cancel a unit in the reduce
function. The update includes Singular 4.4.1 which turned off this behaviour by
default (and included an option, which we shall not use, to turn it on again).
While the behaviour has lead to surprise in experimentation, the main results
have been verified to be unaffected.