Restore exactness of computations
The computation of c_1(ϖ_1) requires significantly more precision than the
others. In addition it is not possible to introduce new coefficients for the
tail part here like it had been done elsewhere, since there is not enough
cancellation. Instead we will have to perform all computations with the full
precision required.